+++ app/src/main/java/com/example/swap_it/ui/chat_list/ChatListScreen.kt
@@ -13,8 +13,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.navigation.NavHostController
-import com.example.swap_it.ui.menu.MenuScreen
-import com.example.swap_it.ui.menu.NavigationModule
+import com.example.swap_it.ui.component.AppBar
+import com.example.swap_it.ui.component.BottomNavigationBar
+
+import com.example.swap_it.ui.navigation.NavigationModule
 
 class ChatListScreen {
     @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter", "NotConstructor")
@@ -23,13 +25,13 @@ class ChatListScreen {
         val navigationModule = NavigationModule()
         Scaffold(
             topBar = {
-                MenuScreen().AppBar(navController = navController)
+                AppBar(navController = navController)
             },
             bottomBar = {
-                navigationModule.BottomNavigationBar(navController)
+                BottomNavigationBar(navController)
             },
 
-        ) {
+            ) {
             Box(
                 modifier = Modifier
                     .fillMaxSize()
@@ -47,3 +49,4 @@ class ChatListScreen {
 
 
     }
+}
