+++ app/src/main/java/com/example/swap_it/ui/navigation/NavigationModule.kt
@@ -7,6 +7,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.example.swap_it.ui.add_item.AddItemScreen
 import com.example.swap_it.ui.chat.ChatListScreen
+import com.example.swap_it.ui.product_list.SearchScreen
 import com.example.swap_it.ui.shopping.Alert
 import com.example.swap_it.ui.shopping.Shopping
 import com.example.swap_it.ui.swap.ShoppingListScreen
@@ -42,6 +43,9 @@ class NavigationModule {
             }
             composable("Alert") {
                 Alert().AlertScreen(modifier = Modifier, navController)
+            }
+            composable("Search") {
+                SearchScreen().SearchScreen(Modifier, navController)
             }
         }
     }
