--- app/src/main/java/com/example/swap_it/ui/product_list/BottomSheet.kt
@@ -1,146 +1,0 @@
-package com.example.swap_it.ui.product_list
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ExperimentalLayoutApi
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material.Divider
-import androidx.compose.material.Icon
-import androidx.compose.material.Text
-import androidx.compose.material.TextButton
-import androidx.compose.material3.SegmentedButton
-import androidx.compose.material3.SegmentedButtonDefaults
-import androidx.compose.material3.SingleChoiceSegmentedButtonRow
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import com.example.swap_it.R
-import com.example.swap_it.ui.component.CategoryButton
-import com.example.swap_it.ui.theme.Gray4
-import com.example.swap_it.ui.theme.Gray5
-import com.example.swap_it.ui.theme.Paddings
-import com.example.swap_it.ui.theme.Primary
-import com.example.swap_it.ui.theme.Typography
-
-class BottomSheet {
-
-    @Composable
-    fun ModalBottomSheetContent() {
-        var option = SortOption.POPULAR
-        Column(modifier = Modifier.padding(Paddings.large)) {
-            Text("정렬", style = Typography.titleLarge)
-            SortButtons { option = it }
-            Text("카테고리", style = Typography.titleLarge)
-            CategoryButtons(option)
-            Spacer(modifier = Modifier.size(48.dp))
-        }
-
-    }
-
-    enum class SortOption(val option: String) {
-        POPULAR("인기순"),
-        RECENT("최신순"),
-        PRICE_HIGH("가격 높은 순"),
-        PRICE_ROW("가격 낮은 순"),
-    }
-
-    @Composable
-    fun SortButtons(
-        initialSortOption: SortOption = SortOption.POPULAR,
-        onSortOptionSelected: (SortOption) -> Unit
-    ) {
-        var selectedOption by remember { mutableStateOf(initialSortOption) }
-        Row(horizontalArrangement = Arrangement.SpaceEvenly, modifier = Modifier.fillMaxWidth()) {
-            SortOption.entries.forEach {
-                SortButton(
-                    text = it.option,
-                    isSelected = selectedOption == it,
-                    onClick = {
-                        selectedOption = it
-                        onSortOptionSelected(it)
-                    }
-                )
-                if (it != SortOption.entries.last()) {
-                    Divider(
-                        modifier = Modifier
-                            .size(1.dp, 16.dp)
-                            .background(Gray4)
-                            .align(Alignment.CenterVertically)
-                    )
-                }
-            }
-        }
-    }
-
-
-    @Composable
-    fun SortButton(
-        text: String,
-        isSelected: Boolean,
-        onClick: () -> Unit
-    ) {
-        val textColor = if (isSelected) Primary else Gray5
-        TextButton(onClick = onClick) {
-            Row {
-                if (isSelected) {
-                    Icon(
-                        painter = painterResource(R.drawable.ic_check),
-                        contentDescription = "선택",
-                        tint = Primary,
-                        modifier = Modifier.size(16.dp)
-                    )
-                }
-                Text(text = text, color = textColor, style = Typography.bodySmall)
-            }
-        }
-    }
-
-    enum class CategoryOption(val categoryName: String) {
-        CAR("자동차"),
-        CLOTH("옷"),
-        MOBILE("모바일 기기"),
-        CUP("컵"),
-        BOOK("책"),
-        ELECTRONIC("가전기기"),
-        FASHION("패션"),
-        FOOD("음식"),
-        ETC("기타"),
-        SNACK("과자"),
-    }
-
-    @OptIn(ExperimentalLayoutApi::class)
-    @Composable
-    fun CategoryButtons(option: SortOption) {
-        FlowRow(
-            horizontalArrangement = Arrangement.Center,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(Paddings.medium),
-
-            ) {
-            CategoryOption.entries.forEach {
-                CategoryButton(
-                    text = it.categoryName,
-                    isSelected = true,
-                    modifier = Modifier.padding(Paddings.small)
-                ) {
-
-                }
-            }
-        }
-    }
-
-}
